pyproject.toml
[tool.poetry]
name = "photoshare-fastapi"
version = "0.1.0"
description = "A FastAPI project for photo sharing"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "photoshare_fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
cloudinary = "^1.41.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
passlib = "^1.7.4"
libgravatar = "^1.0.4"
qrcode = "^7.4.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = {extras = ["dotenv", "email"], version = "^1.10.7"}

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.2"
pytest-mock = "^3.14.0"
sphinx = "^7.4.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
